+++ config/webpack.base.config.js
@@ -4,7 +4,7 @@ const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 const srcPath = path.join(__dirname, '../', 'src', 'docs');
-const paths = require('./build.config.js').paths;
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
 module.exports = {
     target: 'web',
@@ -12,7 +12,7 @@ module.exports = {
     entry: {
         app: path.join(srcPath, 'main.ts'),
         common: [
-            'es6-shim',
+            'core-js',
             'reflect-metadata',
             'zone.js'
         ]
@@ -30,7 +30,7 @@ module.exports = {
 
     module: {
         loaders: [
-            { test: /\.ts$/, loaders: ['ts-loader?transpileOnly=false&configFileName=tsconfig.json', 'angular2-template-loader'] },
+            { test: /\.ts$/, loaders: ['awesome-typescript-loader?transpileOnly=false&configFileName=tsconfig.json', 'angular2-template-loader'] },
             { test: /\.css$/, loader: 'style-loader!raw-loader!autoprefixer-loader' },
             { test: /\.scss$/, loader: 'style-loader!raw-loader!autoprefixer-loader!sass-loader' },
             { test: /\.html/, loader: 'raw-loader' },
@@ -58,7 +58,8 @@ module.exports = {
             inject: true,
             template: path.join(srcPath, 'index.html')
         }),
-        new webpack.NoErrorsPlugin()
+        new webpack.NoErrorsPlugin(),
+        new CheckerPlugin()
     ],
     devtool: 'eval-source-map'
 };