+++ src/components/date-time-picker/date-time-picker-format-provider.service.ts
@@ -1,7 +1,8 @@
 import {Injectable} from '@angular/core';
+import {Observable} from 'rxjs';
 import {DateTimePickerStrings} from './date-time-picker-strings';
 import {defaultStrings} from './date-time-picker-default-strings';
-import {Observable} from 'rxjs';
+import {Moment} from './moment-types';
 
 
 /**
@@ -17,7 +18,7 @@ export class DateTimePickerFormatProvider {
     changed$: Observable<any> = Observable.never();
 
     /** Formats a human-readable string to be displayed in the control input field. */
-    format(date: moment.Moment, displayTime: boolean, displaySeconds: boolean): string {
+    format(date: Moment, displayTime: boolean, displaySeconds: boolean): string {
         let formatString = displayTime ? (displaySeconds ? 'L, LTS' : 'L, LT') : 'L';
         return date.format(formatString);
     }