+++ typings/index.d.ts
@@ -1,7 +1,0 @@
-/// <reference path="globals/es6-shim/index.d.ts" />
-/// <reference path="globals/hammerjs/index.d.ts" />
-/// <reference path="globals/jasmine/index.d.ts" />
-/// <reference path="globals/jquery/index.d.ts" />
-/// <reference path="globals/moment-node/index.d.ts" />
-/// <reference path="globals/sortablejs/index.d.ts" />
-/// <reference path="globals/velocity-animate/index.d.ts" />