+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Gentics UI Core Changelog
 
+## 1.0.1 (in progress)
+
+* (build) Migrate type definitions to @types/*
+* (build) No longer leak type definition dependencies (hammerjs, sortablejs etc.)
+* (build) Add Moment as a peerDependency
+
 ## 1.0.0 (2016-12-06)
 
 ### Breaking Changes